requirements.txt
Django>=3.2.5,<=3.3.0
djangorestframework>=3.12.4,<=3.13.0
psycopg2>=2.9.1,<=2.9.5